LayoutTests/fullscreen/full-screen-enter-while-exiting-mid-completion-handler-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: Fullscreen request aborted by a request to exit fullscreen.
Enter fullscreen with target1
RUN(target1.requestFullscreen())
EVENT(fullscreenchange)
Exit fullscreen with target1
RUN(document.exitFullscreen())
Attempt to enter fullscreen with target2
RUN(target2.requestFullscreen())
EVENT(fullscreenerror)
EVENT(fullscreenchange)
RUN(target2.requestFullscreen())
EVENT(fullscreenchange)
END OF TEST
